+++ Core/src/RendererAPI/ShaderManager.cpp
@@ -1,11 +1,9 @@
 #include "RendererAPI/ShaderManager.h"
-#include "RendererAPI/BufferManager.h"
 
 // libs
 #include <glm/gtc/type_ptr.hpp>
 
 // std
-#include <cstdint>
 #include <fstream>
 #include <sstream>
 #include <iostream>
@@ -13,7 +11,7 @@
 namespace Core
 {
     std::vector<GLuint> ShaderManager::shaders;
-    uint32_t ShaderManager::CreateShader(const char* vertexPath, const char* fragmentPath, const char* geometryPath)
+    size_t ShaderManager::CreateShader(const char* vertexPath, const char* fragmentPath, const char* geometryPath)
     {
         // 1. retrieve the vertex/fragment source code from filePath
         std::string vertexCode;
@@ -137,7 +135,7 @@ namespace Core
         return shaders.size() - 1;
     }
 
-    void ShaderManager::Bind(uint32_t shader)
+    void ShaderManager::Bind(size_t shader)
     {
         glUseProgram(shaders[shader]);
     }
@@ -146,43 +144,43 @@ namespace Core
         glUseProgram(0);
     }
 
-    void ShaderManager::setBool(uint32_t shader, const std::string& name, bool value)
+    void ShaderManager::setBool(size_t shader, const std::string& name, bool value)
     {
         Bind(shader);
         glUniform1i(glGetUniformLocation(shaders[shader], name.c_str()), (int)value);
     }
-    void ShaderManager::setInt(uint32_t shader, const std::string& name, int value)
+    void ShaderManager::setInt(size_t shader, const std::string& name, int value)
     {
         Bind(shader);
         glUniform1i(glGetUniformLocation(shaders[shader], name.c_str()), value);
     }
-    void ShaderManager::setFloat(uint32_t shader, const std::string& name, float value)
+    void ShaderManager::setFloat(size_t shader, const std::string& name, float value)
     {
         Bind(shader);
         glUniform1f(glGetUniformLocation(shaders[shader], name.c_str()), value);
     }
-    void ShaderManager::set2f(uint32_t shader, const std::string& name, float v1, float v2)
+    void ShaderManager::set2f(size_t shader, const std::string& name, float v1, float v2)
     {
         Bind(shader);
         glUniform2f(glGetUniformLocation(shaders[shader], name.c_str()), v1, v2);
     }
-    void ShaderManager::set2f(uint32_t shader, const std::string& name, glm::vec2 v)
+    void ShaderManager::set2f(size_t shader, const std::string& name, glm::vec2 v)
     {
         Bind(shader);
         glUniform2f(glGetUniformLocation(shaders[shader], name.c_str()), v.x, v.y);
     }
-    void ShaderManager::set3f(uint32_t shader, const std::string& name, float v1, float v2, float v3)
+    void ShaderManager::set3f(size_t shader, const std::string& name, float v1, float v2, float v3)
     {
         Bind(shader);
         glUniform3f(glGetUniformLocation(shaders[shader], name.c_str()), v1, v2, v3);
     }
-    void ShaderManager::set3f(uint32_t shader, const std::string& name, glm::vec3 v)
+    void ShaderManager::set3f(size_t shader, const std::string& name, glm::vec3 v)
     {
         Bind(shader);
         glUniform3f(glGetUniformLocation(shaders[shader], name.c_str()), v.x, v.y, v.z);
     }
 
-    void ShaderManager::setmat4(uint32_t shader, const std::string& name, glm::mat4 mat)
+    void ShaderManager::setmat4(size_t shader, const std::string& name, glm::mat4 mat)
     {
         Bind(shader);
         glUniformMatrix4fv(glGetUniformLocation(shaders[shader], name.c_str()), 1, GL_FALSE, glm::value_ptr(mat));
